Move SimpleEvaluator into operators::evaluation and rename to SerialEvaluator.
