+++ src/operators/mod.rs
@@ -12,6 +12,7 @@ use crate::{
 pub mod archive;
 pub mod custom_state;
 pub mod diversity;
+pub mod evaluation;
 pub mod generation;
 pub mod initialization;
 pub mod replacement;
